ipod_fuse: use itdb_filename_on_ipod() and not itdb_resolve_path().
Initially I hadn't noticed itdb_filename_on_ipod(), and thought
itdb_resolve_path() was the only API to obtain the path to a track with
support for case-insensitive filesystems (like VFAT). But this function
was not properly SWIG'ed and was not accessible from Python, so I
created an equivalent my_itdb_resolve_path().

Later I've found out about itdb_filename_on_ipod(), which is available
from the bindings, is somewhat higher-level, and calls itdb_resolve_path()
internally. Win.

For reference, the "Old libgpod, itdb_resolve_path not wrapped correctly"
refers to the patch I sent to the gktpod-devel mailing list to provide a
wrapping:

  http://sourceforge.net/mailarchive/forum.php?thread_name=k2g2bdc975f1004150700q97a15b82h74e6c5af51fea183%40mail.gmail.com&forum_name=gtkpod-devel