ipod_fuse: do not assume "cds" will be available, unbreaking "cdX" creation.
With f8162e5, we fixed the case where cd_nr/cds is 1/1, and there's only
one CD in the album, but we broke the case where there are multiple CDs,
but tracks.cds is not set. We fix this by looking at the number of different
cd_nr tags that are present.